# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "leanamp-backend",
    "version": "0.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3001"
    }
  ],
  "paths": {
...
    "/food/templates": {
      "get": {
        "tags": [
          "food"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "format": "uuid"
                      },
                      "userId": {
                        "type": "string",
                        "format": "uuid"
                      },
                      "createdAt": {
                        "type": "string",
                        "format": "date-time"
                      },
                      "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                      },
                      "name": {
                        "type": "string"
                      },
                      "kcal": {
                        "type": "integer"
                      },
                      "protein": {
                        "type": "number"
                      },
                      "fat": {
                        "type": "number"
                      },
                      "carbs": {
                        "type": "number"
                      },
                      "fiber": {
                        "type": "number"
                      }
                    },
                    "required": [
                      "id",
                      "userId",
                      "createdAt",
                      "updatedAt",
                      "name",
                      "kcal",
                      "protein",
                      "fat",
                      "carbs",
                      "fiber"
                    ],
                    "additionalProperties": false
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "food"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1
                  },
                  "kcal": {
                    "type": "integer",
                    "minimum": 0
                  },
                  "protein": {
                    "type": "number",
                    "minimum": 0
                  },
                  "fat": {
                    "type": "number",
                    "minimum": 0
                  },
                  "carbs": {
                    "type": "number",
                    "minimum": 0
                  },
                  "fiber": {
                    "type": "number",
                    "minimum": 0
                  }
                },
                "required": [
                  "name",
                  "kcal",
                  "protein",
                  "fat",
                  "carbs",
                  "fiber"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FoodTemplate"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/food/templates/{id}": {
      "delete": {
        "tags": [
          "food"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/food/meal-groups": {
      "post": {
        "tags": [
          "food"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "dateISO": {
                    "type": "string",
                    "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
                  },
                  "title": {
                    "type": "string",
                    "minLength": 1
                  },
                  "isExpanded": {
                    "type": "boolean"
                  }
                },
                "required": [
                  "dateISO",
                  "title"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MealGroup"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "food"
        ],
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "format": "uuid"
                      },
                      "userId": {
                        "type": "string",
                        "format": "uuid"
                      },
                      "createdAt": {
                        "type": "string",
                        "format": "date-time"
                      },
                      "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                      },
                      "dateISO": {
                        "type": "string",
                        "format": "date-time"
                      },
                      "title": {
                        "type": "string"
                      },
                      "isExpanded": {
                        "type": "boolean"
                      }
                    },
                    "required": [
                      "id",
                      "userId",
                      "createdAt",
                      "updatedAt",
                      "dateISO",
                      "title",
                      "isExpanded"
                    ],
                    "additionalProperties": false
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/food/meal-groups/{id}": {
      "patch": {
        "tags": [
          "food"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "minLength": 1
                  },
                  "isExpanded": {
                    "type": "boolean"
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MealGroup"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/food/entries": {
      "get": {
        "tags": [
          "food"
        ],
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "format": "uuid"
                      },
                      "userId": {
                        "type": "string",
                        "format": "uuid"
                      },
                      "createdAt": {
                        "type": "string",
                        "format": "date-time"
                      },
                      "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                      },
                      "dateISO": {
                        "type": "string",
                        "format": "date-time"
                      },
                      "name": {
                        "type": "string"
                      },
                      "kcal": {
                        "type": "integer"
                      },
                      "protein": {
                        "type": "number"
                      },
                      "fat": {
                        "type": "number"
                      },
                      "carbs": {
                        "type": "number"
                      },
                      "fiber": {
                        "type": "number"
                      },
                      "multiplier": {
                        "type": "number"
                      },
                      "type": {
                        "type": "string",
                        "enum": [
                          "quick_add",
                          "manual",
                          "photo",
                          "ai_text"
                        ]
                      },
                      "groupId": {
                        "type": "string",
                        "format": "uuid",
                        "nullable": true
                      },
                      "note": {
                        "type": "string",
                        "nullable": true
                      }
                    },
                    "required": [
                      "id",
                      "userId",
                      "createdAt",
                      "updatedAt",
                      "dateISO",
                      "name",
                      "kcal",
                      "protein",
                      "fat",
                      "carbs",
                      "fiber",
                      "multiplier",
                      "type",
                      "groupId",
                      "note"
                    ],
                    "additionalProperties": false
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "food"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "dateISO": {
                    "type": "string",
                    "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
                  },
                  "name": {
                    "type": "string",
                    "minLength": 1
                  },
                  "kcal": {
                    "type": "integer",
                    "minimum": 0
                  },
                  "protein": {
                    "type": "number",
                    "minimum": 0
                  },
                  "fat": {
                    "type": "number",
                    "minimum": 0
                  },
                  "carbs": {
                    "type": "number",
                    "minimum": 0
                  },
                  "fiber": {
                    "type": "number",
                    "minimum": 0
                  },
                  "multiplier": {
                    "type": "number",
                    "exclusiveMinimum": true,
                    "minimum": 0
                  },
                  "type": {
                    "type": "string",
                    "enum": [
                      "quick_add",
                      "manual",
                      "photo",
                      "ai_text"
                    ]
                  },
                  "groupId": {
                    "type": "string",
                    "format": "uuid",
                    "nullable": true
                  },
                  "note": {
                    "type": "string",
                    "nullable": true
                  }
                },
                "required": [
                  "dateISO",
                  "name",
                  "kcal",
                  "protein",
                  "fat",
                  "carbs",
                  "fiber",
                  "type"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FoodEntry"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/food/entries/{id}": {
      "patch": {
        "tags": [
          "food"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "dateISO": {
                    "type": "string",
                    "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
                  },
                  "name": {
                    "type": "string",
                    "minLength": 1
                  },
                  "kcal": {
                    "type": "integer",
                    "minimum": 0
                  },
                  "protein": {
                    "type": "number",
                    "minimum": 0
                  },
                  "fat": {
                    "type": "number",
                    "minimum": 0
                  },
                  "carbs": {
                    "type": "number",
                    "minimum": 0
                  },
                  "fiber": {
                    "type": "number",
                    "minimum": 0
                  },
                  "multiplier": {
                    "type": "number",
                    "exclusiveMinimum": true,
                    "minimum": 0
                  },
                  "type": {
                    "type": "string",
                    "enum": [
                      "quick_add",
                      "manual",
                      "photo",
                      "ai_text"
                    ]
                  },
                  "groupId": {
                    "type": "string",
                    "format": "uuid",
                    "nullable": true
                  },
                  "note": {
                    "type": "string",
                    "nullable": true
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FoodEntry"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "food"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "details": {}
            },
            "required": [
              "code",
              "message"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false
      },
...
      "FoodTemplate": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "name": {
            "type": "string"
          },
          "kcal": {
            "type": "integer"
          },
          "protein": {
            "type": "number"
          },
          "fat": {
            "type": "number"
          },
          "carbs": {
            "type": "number"
          },
          "fiber": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "userId",
          "createdAt",
          "updatedAt",
          "name",
          "kcal",
          "protein",
          "fat",
          "carbs",
          "fiber"
        ],
        "additionalProperties": false
      },
      "MealGroup": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "dateISO": {
            "type": "string",
            "format": "date-time"
          },
          "title": {
            "type": "string"
          },
          "isExpanded": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "userId",
          "createdAt",
          "updatedAt",
          "dateISO",
          "title",
          "isExpanded"
        ],
        "additionalProperties": false
      },
      "FoodEntry": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "dateISO": {
            "type": "string",
            "format": "date-time"
          },
          "name": {
            "type": "string"
          },
          "kcal": {
            "type": "integer"
          },
          "protein": {
            "type": "number"
          },
          "fat": {
            "type": "number"
          },
          "carbs": {
            "type": "number"
          },
          "fiber": {
            "type": "number"
          },
          "multiplier": {
            "type": "number"
          },
          "type": {
            "type": "string",
            "enum": [
              "quick_add",
              "manual",
              "photo",
              "ai_text"
            ]
          },
          "groupId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "note": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "userId",
          "createdAt",
          "updatedAt",
          "dateISO",
          "name",
          "kcal",
          "protein",
          "fat",
          "carbs",
          "fiber",
          "multiplier",
          "type",
          "groupId",
          "note"
        ],
        "additionalProperties": false
      },
...
    }
  }
}